requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==21.2.0

chromadb==0.4.24
numpy<2.0

huggingface-hub==0.23.4
python-dotenv==1.0.1
